Update and rename Game.java to GameLogic.java
