+++ GameLogic.java
@@ -1,0 +1,286 @@
+import java.util.*;import java.util.*;
+
+public class GameLogic {
+
+    private static boolean manual;
+    private final int CLASSES = 4;
+	private final int SHIPS = 10;
+	private final int NUM_OF_PLAYERS = 2;
+
+
+	private Cell[][] field = new Cell[10][10];
+	private boolean automode;
+
+
+	private ArrayList<Ship> getShips() {
+		return ships;
+	}
+
+	private ArrayList<Ship> ships = new ArrayList<Ship>();
+	
+	private ArrayList<Player> players = new ArrayList<Player>();
+	
+	private Display d = new Display();
+	
+	
+	
+	private Cell[][] createField(){
+		for(int i=0; i<field.length; i++){
+			
+			for(int j=0; j<field[i].length; j++){
+				field [i][j] = Cell.EMPTY;
+			}
+			
+		}
+		return field;
+	}
+	
+	private void printField(Cell[][] field1, Cell[][] field2){
+	    d.printField(field1, field2);
+	}
+	
+	/*public void drawBoard(){
+		System.out.print(" " + "\t");
+		for(int h=0; h<heading.length; h++){
+		System.out.print(heading[h]);
+		}
+		System.out.println();
+		for(int i=0; i<field.length; i++){
+			System.out.print(i+1 + "\t");
+			for(int j=0; j<field[i].length; j++){
+				System.out.print(field[i][j]);
+			}
+			System.out.println("|");
+		}
+	}
+	*/
+	
+
+	
+	
+	private void shot(){
+		// move to Display
+//		System.out.println("Shoot!");
+//		Scanner s = new Scanner(System.in);
+//		String shot = s.nextLine();
+//		shot = shot.replaceAll("[Aa]", "0");
+//		shot = shot.replaceAll("[Bb]", "1");
+//		shot = shot.replaceAll("[Cc]", "2");
+//		shot = shot.replaceAll("[Dd]", "3");
+//		shot = shot.replaceAll("[Ee]", "4");
+//		shot = shot.replaceAll("[Ff]", "5");
+//		shot = shot.replaceAll("[Gg]", "6");
+//		shot = shot.replaceAll("[Hh]", "7");
+//		shot = shot.replaceAll("[Ii]", "8");
+//		shot = shot.replaceAll("[Jj]", "9");
+//		//char[] sh = shot.toCharArraY();
+//		Cell c = new Cell("Miss");
+//		System.out.println(shot);
+//	  field[(Character.getNumericValue(shot.charAt(1))-1)][Character.getNumericValue(shot.charAt(0))] = c.getShape();
+//
+	}
+
+	private int generateRandom(int b){
+		return new Random().nextInt(b);
+	}
+
+	private  void autoShoot(){
+		d.autoShoot();
+        int xToShoot = generateRandom(field.length);
+        int yToShoot = generateRandom(field[xToShoot].length);
+        if(field[xToShoot][yToShoot]==Cell.DECK){
+        	field[xToShoot][yToShoot] = Cell.HIT;
+		} else {field[yToShoot][xToShoot] = Cell.MISS;}
+		System.out.println("shot at :" + d.getXCoord(xToShoot) + ", " + d.getYCoord(yToShoot));
+    }
+
+	private void salute(){
+
+		d.salute();
+	}
+		
+		public void endGame(){
+			d.weHaveWinner(players.get(0).getName());
+		}
+	
+	// Start
+	public static void main(String[] args){
+		GameLogic g = new GameLogic();
+
+		g.salute();
+		Player playerOne = new Player();
+		playerOne.setName(g.getPlayerName());
+		g.players.add(playerOne);
+		g.requestMode();
+		Player playerTwo = new Player();
+		if(!g.automode){
+		playerTwo.setName(g.getPlayerName());}
+		else {playerTwo.setName("Computer");}
+		g.players.add(playerTwo);
+		g.players.get(0).setPlayerField(g.createField());
+		g.players.get(1).setPlayerField(g.createField());
+
+
+		g.printField(g.players.get(0).getField(), g.players.get(1).getField());
+		manual = false;
+        if (manual) {
+            playerOne.makeShipsManually();
+        } else {playerOne.fillField();
+        }
+        g.printField(playerOne.getField(), playerTwo.getField());
+//		for(Ship i: g.getShips()){
+//			System.out.println("Ship 1");
+//			for(Deck d: i.getDecksList()){
+//				System.out.println(d.getX()+1 + " " + d.getY()+1);
+//			}
+//		}
+		for (int i = 0; i<4;i++) {
+			g.autoShoot();
+			g.printField(g.field, g.players.get(1).getField());
+		}
+		g.printField(g.field, g.players.get(1).getField());
+	}
+
+	private void requestMode() {
+		automode = d.requestMode();
+
+	}
+
+	private String getPlayerName() {
+		return d.getPlayerName();
+	}
+}
+
+
+
+
+import java.io.*;
+
+public class Game {
+
+	// Game constants
+	private final int ONEDECK_SHIPS = 4;
+	private final int TWODECK_SHIPS = 3;
+	private final int THREEDECK_SHIPS = 2;
+	private final int FOURDECK_SHIPS = 1;
+	private final int ONE_DECK = 1;
+	private final int TWO_DECK = 2;
+	private final int THREE_DECK = 3;
+	private final int FOUR_DECK = 4;
+	private final int CLASSES = 4;
+	private final int SHIPS = 10;
+	private final int NUM_OF_PLAYERS = 2;
+	
+
+	String[][] field = new String[10][10];
+	ArrayList<Ship> ships = new ArrayList<Ship>();
+	
+	ArrayList<Player> players = new ArrayList<Player>();
+	
+	public Display d = new Display();
+	
+	
+	
+	public void createField(){
+		for(int i=0; i<field.length; i++){
+			
+			for(int j=0; j<field[i].length; j++){
+				Cell c = new Cell("Empty");
+				field [i][j] = c.getShape();
+			}
+			
+		}
+	}
+	
+	public void printField(String[][] field){
+	    d.printField(field);
+	}
+	
+	/*public void drawBoard(){
+		System.out.print(" " + "\t");
+		for(int h=0; h<heading.length; h++){
+		System.out.print(heading[h]);
+		}
+		System.out.println();
+		for(int i=0; i<field.length; i++){
+			System.out.print(i+1 + "\t");
+			for(int j=0; j<field[i].length; j++){
+				System.out.print(field[i][j]);
+			}
+			System.out.println("|");
+		}
+	}
+	*/
+	
+	public void fillField(){
+		
+		makeShips(ONEDECK_SHIPS, ONE_DECK);
+		makeShips(TWODECK_SHIPS, TWO_DECK);
+		makeShips(THREEDECK_SHIPS, THREE_DECK);
+		makeShips(FOURDECK_SHIPS, FOUR_DECK);
+		
+		d.shipsAreReady();
+	}
+	
+	public void makeShips(int class, int decks){
+		for(int i=0; i<class; i++){
+		Ship s = new Ship(decks, new Random().nextInt(field.length), new Random().nextInt(field[1].length));
+		Cell c = new Cell("Deck");
+		field[s.decksList.get(0).getX()][s.decksList.get(0).getY()] = c.getShape();
+			ships.add(s);
+		}
+	}
+	
+	
+	public void shot(){
+		System.out.println("Shoot!");
+		Scanner s = new Scanner(System.in);
+		String shot = s.nextLine();
+		shot = shot.replaceAll("[Aa]", "0");
+		shot = shot.replaceAll("[Bb]", "1");
+		shot = shot.replaceAll("[Cc]", "2");
+		shot = shot.replaceAll("[Dd]", "3");
+		shot = shot.replaceAll("[Ee]", "4");
+		shot = shot.replaceAll("[Ff]", "5");
+		shot = shot.replaceAll("[Gg]", "6");
+		shot = shot.replaceAll("[Hh]", "7");
+		shot = shot.replaceAll("[Ii]", "8");
+		shot = shot.replaceAll("[Jj]", "9");
+		//char[] sh = shot.toCharArraY();
+		Cell c = new Cell("Miss");
+		System.out.println(shot);
+	  field[(Character.getNumericValue(shot.charAt(1))-1)][Character.getNumericValue(shot.charAt(0))] = c.getShape();
+		
+	}
+
+	public void salute(){
+
+		d.salute();
+	}
+		
+		public void endGame(){
+			d.weHaveWinner(players.get(0).getName());
+		}
+	
+	// Start
+	public static void main(String[] args){
+		Game g = new Game();
+
+		g.salute();
+		Player playerOne = new Player();
+		playerOne.setName(g.getPlayerName());
+		g.createField();
+		g.printField(g.field);
+		g.fillField();
+		g.printField(g.field);
+		g.shot();
+		g.printField(g.field);
+	}
+
+	private String getPlayerName() {
+		return d.getPlayerName();
+	}
+}
+
+
+
